feat(api): write API to recommend users based on their location

+++ UserService/src/main/java/com/example/UserService/user/dto/request/RequestCreate.java
@@ -37,7 +37,7 @@ public class RequestCreate {
     String faculty;
     String phone;
     Date dob;
-    Long latitude;
-    Long longitude;
+    Double latitude;
+    Double longitude;
     String jsonSchedule;
 }
